requirements.txt
fastapi>=0.99.0
uvicorn
sqlalchemy
asyncpg
pydantic
python-jose[cryptography]
passlib[bcrypt]
bcrypt>=3.1.0,<4.1.0
gunicorn
python-multipart